feat: make map landing page, mv current to about (#907)
* feat: make map landing page, mv current to about

* fix: update copyright year

* feat: improve hosted text

* fix: copyright year, twitter icon

* fix: translations

* fix: openSenseLab logo

* fix: rm scroll snapping

* fix: translations

* feat: animate heading

* fix: rm features

* feat: show copyright year dynamically

---------
