+++ app/components/header/menu/index.tsx
@@ -6,6 +6,7 @@ import {
 	Compass,
 	PlusIcon,
 	DownloadIcon,
+	Info,
 } from 'lucide-react'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
@@ -104,7 +105,13 @@ export default function Menu({ devices }: MenuProps) {
 									</div>
 								)}
 
-								{!isExplore && (
+								<Link to="/about">
+									<DropdownMenuItem className="cursor-pointer">
+										<Info className="mr-2 h-5 w-5" />
+										<span>{t('about_label')}</span>
+									</DropdownMenuItem>
+							</Link>
+							{!isExplore && (
 									<Link to="/explore">
 										<DropdownMenuItem className="cursor-pointer">
 											<Compass className="mr-2 h-5 w-5" />
