+++ public/locales/de/header.json
@@ -1,5 +1,5 @@
 {
-	"Explore": "Erkunden",
+	"Map": "Karte",
 	"Features": "Features",
 	"Tools": "Werkzeuge",
 	"Use Cases": "Anwendungen",
