+++ public/locales/de/landing.json
@@ -1,5 +1,5 @@
 {
 	"introduction": "Die wahrscheinlich größte offene Plattform für Umwelt- und Sensordaten weltweit ermöglicht es allen Menschen kostenfrei auf umfangreiche Daten zuzugreifen, diese zu analysieren und eigene Messungen zu veröffentlichen. Dadurch ist ein einzigartiger Datensatz mit Echtzeitdaten überall auf der Welt entstanden, der nachprüfbare lokale wie globale Fakten zu Umweltphänomenen wie dem Klimawandel liefert.",
 	"donate": "Spenden",
-	"explore": "Erkunden"
+	"map": "Karte"
 }
