+++ public/locales/de/menu.json
@@ -1,27 +1,28 @@
 {
-	"title": "Wilkommen",
-	"subtitle": "Bitte loggen Sie sich ein, um mehr Inhalte zu sehen.",
-	"explore_label": "Erkunden",
-	"profile_label": "Profil",
-	"settings_label": "Einstellungen",
-	"my_devices_label": "Meine Geräte",
+  "title": "Wilkommen",
+  "subtitle": "Bitte loggen Sie sich ein, um mehr Inhalte zu sehen.",
+  "explore_label": "Erkunden",
+  "profile_label": "Profil",
+  "settings_label": "Einstellungen",
+  "about_label": "Über OpenSenseMap",
+  "my_devices_label": "Meine Geräte",
 	"download_label": "Datendownload",
-	"add_device_label": "Gerät hinzufügen",
-	"tutorials_label": "Tutorials",
-	"api_docs_label": "API Dokumentation",
-	"faq_label": "FAQ",
-	"contact_label": "Kontakt",
-	"imprint_label": "Impressum",
-	"data_protection_label": "Datenschutz",
-	"donate_label": "Spenden",
-	"tos": "Nutzungsbedingungen",
-	"promotion_label": "Förderung",
-	"login_label": "Einloggen",
-	"logout_label": "Ausloggen",
-	"toast_login_success": "Erfolgreich eingeloggt",
-	"toast_logout_success": "Erfolgreich ausgeloggt",
-	"toast_user_creation_success": "Benutzer erfolgreich erstellt",
-	"community_label": "Community Forum",
+  "add_device_label": "Gerät hinzufügen",
+  "tutorials_label": "Tutorials",
+  "api_docs_label": "API Dokumentation",
+  "faq_label": "FAQ",
+  "contact_label": "Kontakt",
+  "imprint_label": "Impressum",
+  "data_protection_label": "Datenschutz",
+  "donate_label": "Spenden",
+  "tos": "Nutzungsbedingungen",
+  "promotion_label": "Förderung",
+  "login_label": "Einloggen",
+  "logout_label": "Ausloggen",
+  "toast_login_success": "Erfolgreich eingeloggt",
+  "toast_logout_success": "Erfolgreich ausgeloggt",
+  "toast_user_creation_success": "Benutzer erfolgreich erstellt",
+  "community_label": "Community Forum",
 	"devices": "Geräte",
 	"measurements": "Messungen"
 }
