+++ public/locales/de/partners.json
@@ -1,4 +1,4 @@
 {
 	"Partners": "Partner",
-	"hosted": "gehostet von"
+	"made_by": "Made with love in Münster by"
 }
