+++ public/locales/en/header.json
@@ -1,5 +1,5 @@
 {
-	"Explore": "Explore",
+	"Map": "Map",
 	"Features": "Features",
 	"Tools": "Tools",
 	"Use Cases": "Use Cases",
