+++ public/locales/en/landing.json
@@ -1,5 +1,5 @@
 {
 	"introduction": "Probably the largest open platform for environmental and sensor data in the world, it allows anyone to access extensive data free of charge, analyze it and publish their own measurements. This has created a unique dataset with real-time data all over the world, providing verifiable local as well as global facts on environmental phenomena such as climate change.",
 	"donate": "Donate",
-	"explore": "Explore"
+	"map": "Map"
 }
