+++ public/locales/en/menu.json
@@ -1,27 +1,28 @@
 {
-	"title": "Welcome",
-	"subtitle": "Please sign in to see more content",
-	"explore_label": "Explore",
-	"profile_label": "Profile",
-	"settings_label": "Settings",
+  "title": "Welcome",
+  "subtitle": "Please sign in to see more content",
+  "explore_label": "Explore",
+  "profile_label": "Profile",
+  "about_label": "About",
+  "settings_label": "Settings",
 	"download_label": "Download data",
-	"my_devices_label": "My Devices",
-	"add_device_label": "Add Device",
-	"tutorials_label": "Tutorials",
-	"api_docs_label": "API Docs",
-	"faq_label": "FAQ",
-	"contact_label": "Contact",
-	"imprint_label": "Imprint",
-	"data_protection_label": "Data Protection",
-	"donate_label": "Donate",
-	"tos": "Terms of Service",
-	"promotion_label": "Promotion",
-	"login_label": "Log in",
-	"logout_label": "Log out",
-	"toast_login_success": "Successfully logged in",
-	"toast_logout_success": "Successfully logged out",
-	"toast_user_creation_success": "Successfully created account",
-	"community_label": "Community Forum",
+  "my_devices_label": "My Devices",
+  "add_device_label": "Add Device",
+  "tutorials_label": "Tutorials",
+  "api_docs_label": "API Docs",
+  "faq_label": "FAQ",
+  "contact_label": "Contact",
+  "imprint_label": "Imprint",
+  "data_protection_label": "Data Protection",
+  "donate_label": "Donate",
+  "tos": "Terms of Service",
+  "promotion_label": "Promotion",
+  "login_label": "Log in",
+  "logout_label": "Log out",
+  "toast_login_success": "Successfully logged in",
+  "toast_logout_success": "Successfully logged out",
+  "toast_user_creation_success": "Successfully created account",
+  "community_label": "Community Forum",
 	"devices": "Devices",
 	"measurements": "Measurements"
 }
