+++ public/locales/en/partners.json
@@ -1,4 +1,4 @@
 {
 	"Partners": "Partners",
-	"hosted": "hosted by"
+	"made_by": "Made with love in Münster by"
 }
